+++ grails-app/controllers/UrlMappings.groovy
@@ -1,4 +1,4 @@
-package com.miniasaaaslw.controller
+
 
 class UrlMappings {
